Update component tree
