+++ src/components/documentParser/InvoiceParser/InvoiceDetailsViewer.tsx
@@ -1,4 +1,4 @@
-import { Button, Card, CardContent, Stack } from '@mui/material';
+import { Card, CardContent } from '@mui/material';
 import { DataGrid, GridColumns } from '@mui/x-data-grid';
 import { FC } from 'react';
 import { Invoice } from 'src/@types/invoiceParsers';
@@ -15,23 +15,16 @@ const columns: GridColumns<Invoice> = [
 ];
 interface InvoiceDetailsViewerProps {
   data: Invoice[];
-  onReset: VoidFunction;
 }
-export const InvoiceDetailsViewer: FC<InvoiceDetailsViewerProps> = ({ data, onReset }) => (
-  <Stack>
-    <Button sx={{ alignSelf: 'end' }} onClick={onReset}>
-      Reset
-    </Button>
-
-    <Card>
-      <CardContent>
-        <DataGrid
-          autoHeight
-          columns={columns}
-          rows={data}
-          getRowId={(row) => row.infoTributaria.claveAcceso}
-        />
-      </CardContent>
-    </Card>
-  </Stack>
+export const InvoiceDetailsViewer: FC<InvoiceDetailsViewerProps> = ({ data }) => (
+  <Card>
+    <CardContent>
+      <DataGrid
+        autoHeight
+        columns={columns}
+        rows={data}
+        getRowId={(row) => row.infoTributaria.claveAcceso}
+      />
+    </CardContent>
+  </Card>
 );
